Add RetryingMetaStoreClientFactory
