+++ paimon-hive/paimon-hive-catalog/src/main/java/org/apache/paimon/hive/HiveCatalog.java
@@ -37,15 +37,10 @@ import org.apache.paimon.table.TableType;
 import org.apache.paimon.types.DataField;
 import org.apache.paimon.types.DataTypes;
 
-import org.apache.paimon.shade.guava30.com.google.common.collect.ImmutableMap;
-
 import org.apache.flink.table.hive.LegacyHiveClasses;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.hive.conf.HiveConf;
-import org.apache.hadoop.hive.metastore.HiveMetaHookLoader;
-import org.apache.hadoop.hive.metastore.HiveMetaStoreClient;
 import org.apache.hadoop.hive.metastore.IMetaStoreClient;
-import org.apache.hadoop.hive.metastore.RetryingMetaStoreClient;
 import org.apache.hadoop.hive.metastore.api.Database;
 import org.apache.hadoop.hive.metastore.api.FieldSchema;
 import org.apache.hadoop.hive.metastore.api.NoSuchObjectException;
@@ -63,20 +58,15 @@ import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.InputStream;
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
-import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
-import java.util.Map.Entry;
 import java.util.Optional;
 import java.util.Set;
-import java.util.concurrent.ConcurrentHashMap;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
@@ -555,114 +545,8 @@ public class HiveCatalog extends AbstractCatalog {
         return Lock.fromCatalog(lock, identifier);
     }
 
-    private static final Map<Class<?>[], HiveMetastoreProxySupplier> PROXY_SUPPLIERS =
-            ImmutableMap.<Class<?>[], HiveMetastoreProxySupplier>builder()
-                    // for hive 1.x
-                    .put(
-                            new Class<?>[] {
-                                HiveConf.class,
-                                HiveMetaHookLoader.class,
-                                ConcurrentHashMap.class,
-                                String.class
-                            },
-                            (getProxyMethod, hiveConf, clientClassName) ->
-                                    (IMetaStoreClient)
-                                            getProxyMethod.invoke(
-                                                    null,
-                                                    hiveConf,
-                                                    (HiveMetaHookLoader) (tbl -> null),
-                                                    new ConcurrentHashMap<>(),
-                                                    clientClassName))
-                    // for hive 2.x
-                    .put(
-                            new Class<?>[] {
-                                HiveConf.class,
-                                HiveMetaHookLoader.class,
-                                ConcurrentHashMap.class,
-                                String.class,
-                                Boolean.TYPE
-                            },
-                            (getProxyMethod, hiveConf, clientClassName) ->
-                                    (IMetaStoreClient)
-                                            getProxyMethod.invoke(
-                                                    null,
-                                                    hiveConf,
-                                                    (HiveMetaHookLoader) (tbl -> null),
-                                                    new ConcurrentHashMap<>(),
-                                                    clientClassName,
-                                                    true))
-                    // for hive 3.x
-                    .put(
-                            new Class<?>[] {
-                                Configuration.class,
-                                HiveMetaHookLoader.class,
-                                ConcurrentHashMap.class,
-                                String.class,
-                                Boolean.TYPE
-                            },
-                            (getProxyMethod, hiveConf, clientClassName) ->
-                                    (IMetaStoreClient)
-                                            getProxyMethod.invoke(
-                                                    null,
-                                                    hiveConf,
-                                                    (HiveMetaHookLoader) (tbl -> null),
-                                                    new ConcurrentHashMap<>(),
-                                                    clientClassName,
-                                                    true))
-                    .build();
-    // If clientClassName is HiveMetaStoreClient,
-    // we can revert to the simplest creation method,
-    // which allows us to use shaded Hive packages to avoid dependency conflicts,
-    // such as using apache-hive2.jar in Presto and Trino.
-    private static final Map<Class<?>[], HiveMetastoreProxySupplier> PROXY_SUPPLIERS_SHADED =
-            ImmutableMap.<Class<?>[], HiveMetastoreProxySupplier>builder()
-                    .put(
-                            new Class<?>[] {HiveConf.class},
-                            (getProxyMethod, hiveConf, clientClassName) ->
-                                    (IMetaStoreClient) getProxyMethod.invoke(null, hiveConf))
-                    .put(
-                            new Class<?>[] {HiveConf.class, Boolean.TYPE},
-                            (getProxyMethod, hiveConf, clientClassName) ->
-                                    (IMetaStoreClient) getProxyMethod.invoke(null, hiveConf, true))
-                    .put(
-                            new Class<?>[] {Configuration.class, Boolean.TYPE},
-                            (getProxyMethod, hiveConf, clientClassName) ->
-                                    (IMetaStoreClient) getProxyMethod.invoke(null, hiveConf, true))
-                    .build();
-
     static IMetaStoreClient createClient(HiveConf hiveConf, String clientClassName) {
-        Method getProxy = null;
-        HiveMetastoreProxySupplier supplier = null;
-        RuntimeException methodNotFound =
-                new RuntimeException(
-                        "Failed to find desired getProxy method from RetryingMetaStoreClient");
-        Map<Class<?>[], HiveMetastoreProxySupplier> suppliers =
-                new LinkedHashMap<>(PROXY_SUPPLIERS);
-        if (HiveMetaStoreClient.class.getName().equals(clientClassName)) {
-            suppliers.putAll(PROXY_SUPPLIERS_SHADED);
-        }
-        for (Entry<Class<?>[], HiveMetastoreProxySupplier> entry : suppliers.entrySet()) {
-            Class<?>[] classes = entry.getKey();
-            try {
-                getProxy = RetryingMetaStoreClient.class.getMethod("getProxy", classes);
-                supplier = entry.getValue();
-            } catch (NoSuchMethodException e) {
-                methodNotFound.addSuppressed(e);
-            }
-        }
-        if (getProxy == null) {
-            throw methodNotFound;
-        }
-
-        IMetaStoreClient client;
-        try {
-            client = supplier.get(getProxy, hiveConf, clientClassName);
-        } catch (Exception e) {
-            throw new RuntimeException(e);
-        }
-        return isNullOrWhitespaceOnly(hiveConf.get(HiveConf.ConfVars.METASTOREURIS.varname))
-                ? client
-                : HiveMetaStoreClient.newSynchronizedClient(client);
+        return new RetryingMetaStoreClientFactory().createClient(hiveConf, clientClassName);
     }
 
     public static HiveConf createHiveConf(
@@ -785,11 +669,5 @@ public class HiveCatalog extends AbstractCatalog {
 
     public static String possibleHiveConfPath() {
         return System.getenv("HIVE_CONF_DIR");
-    }
-
-    /** Function interface for creating hive metastore proxy. */
-    public interface HiveMetastoreProxySupplier {
-        IMetaStoreClient get(Method getProxyMethod, Configuration conf, String clientClassName)
-                throws IllegalAccessException, IllegalArgumentException, InvocationTargetException;
     }
 }
